pyproject.toml
[project]
name = "logos"
version = "0.1.0"
description = "On-brand QR code and asset generator for Cascading Labs projects"
requires-python = ">=3.11"
dependencies = [
    "segno>=1.6",
    "pillow>=10.0",
]
